+++ src/main/java/com/nhnacademy/bookpubshop/author/service/AuthorService.java
@@ -1,7 +1,8 @@
 package com.nhnacademy.bookpubshop.author.service;
 
-import com.nhnacademy.bookpubshop.author.dto.CreateAuthorRequestDto;
-import com.nhnacademy.bookpubshop.author.dto.GetAuthorResponseDto;
+import com.nhnacademy.bookpubshop.author.dto.request.CreateAuthorRequestDto;
+import com.nhnacademy.bookpubshop.author.dto.request.ModifyAuthorRequestDto;
+import com.nhnacademy.bookpubshop.author.dto.response.GetAuthorResponseDto;
 import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,9 +18,16 @@ public interface AuthorService {
      * 저자를 생성하는 메서드입니다.
      *
      * @param author 저자를 받습니다.
-     * @return 생성한 저자를 반환합니다.
      */
-    GetAuthorResponseDto createAuthor(CreateAuthorRequestDto author);
+    void createAuthor(CreateAuthorRequestDto author);
+
+    /**
+     * 저자를 수정하는 서비스.
+     *
+     * @param authorNo 저자 번호
+     * @param dto      수정할 저자
+     */
+    void modifyAuthorName(Integer authorNo, ModifyAuthorRequestDto dto);
 
     /**
      * 모든 저자를 반환하는 메서드입니다.
